requirements.txt
numpy~=1.20.3
tensorflow-gpu~=2.1.0
matplotlib~=3.3.4
ipython~=7.22.0
pandas~=1.2.4
pyyaml~=5.4.1
opencv-python~=4.5.2